update
Menu closing fully functional.

Strangely, despite the menu opening and scrolling never causes black bars which that write all its tiles in 1 frame, closing the menu, using !Freeram_CustomL3Menu_WritePhase so that the tiles to write are split into multiple frames, DO causes black bars on the frame of clearing out the options (after removing the cursor but before the scroll arrows). Not sure why is this happening, is it the RLE more prone to do this?