optimization
made it so that the tables only contain tile numbers for strings, because every tile properties is a value of $38 (%00111000), save for the OK and CANCEL options (which is currently removed but will be re-added later). This also makes it easy to edit each line of options. I again made use of the ASCII that asar uses to convert a string into a table.