Fully working, but haven't made code to check the string for password usage yet...
I made the OK and CANCEL options 1-tile, as this would simplify how to handle cursor movement to those options. I also added a space character option so that you don't need to press R, which by the way, the R button doesn't overwrite the characters the cursor is on unlike the space, allowing intuitive UX string input that have spaces.